pyproject.toml
[project]
name = "manatal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alive-progress>=3.2.0",
    "fastapi>=0.115.13",
    "google-genai>=1.20.0",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.14.3",
    "uvicorn>=0.34.3",
]
